+++ README.md
@@ -1,1 +1,9 @@
-# games-rating
+# Привет
+
+Это стартовый архив для второй проектной работы по бэкенду.
+
+## Что здесь?
+
+Папка `public` — клиентская сторона всего приложения. Эту небольшую вёрстку серверное приложение будет раздавать, чтобы можно было перейти по адресу `localhost:3005`, поиграть в любую игру и оценить её, указав соответствующий рейтинг.
+
+## Удачи, ты точно справишься
