+++ appModules/rating/calculations.js
@@ -11,4 +11,5 @@ function updateRating(array, id, rating) {
     array[index].rating += rating;
     return array;
 }
+
 module.exports = { createRating, updateRating };
