+++ appModules/rating/index.js
@@ -1,6 +1,7 @@
 const config = require("./config");
 const makeRatingFile = require("./rating-file");
-const { updateRating, createRating } = require("./calculations");
+const { createRating, updateRating } = require("./calculations");
+
 
 module.exports = {
     config,
