+++ appModules/rating/rating-file.js
@@ -1,32 +1,7 @@
-const { randomFill } = require("crypto");
-
 const fs = require("fs").promises;
 
-function makeRatingFile(path, array){
-    fs.readFile(path, "utf8", (err, ratingFile) => {
-    if (err){
-        console.log(err);
-        return;
-    }
-
-    const ratingArray = JSON.parse(ratingFile);
-
-    array.array.forEach(item => {
-        if (!ratingArray.find((el) => el.id === item.id)){
-            let obj ={
-                id: item.id,
-                title: item.title,
-                image: item.image, 
-                link: item.link,
-                description: item.description,
-                rating: 0, 
-            };
-            ratingArray.push(obj);
-        }
-    });
-    fs.writeFile(path, JSON.stringify(ratingArray), () =>{
-        console.log("Файл записа!");
-    });
-    });
+async function makeRatingFile(path, array){
+    
 }
+
 module.exports = makeRatingFile;
