+++ controllers/game.js
@@ -1,5 +1,5 @@
 const fs = require("fs").promises;
-const { getRandomGame, config } = require("../appModules/api");
+const { config } = require("../appModules/api");
 const sendAllGames = (req, res) => {
   res.send(req.games);
 };
@@ -10,18 +10,5 @@ const sendUpdatedGames = (req, res) => {
     updated: req.updatedObject
   });
 };
-
-async function gameRouteController(res) {
-  try {
-    const ratingFile = await fs.readFile(config.PATH_TO_RATING_FILE);
-    const data = JSON.parse(ratingFile);
-    const game = getRandomGame(data); // Получаем случайную игру
-    res.setHeader("Content-Type", "application/json");
-    res.end(JSON.stringify(game));
-  } catch (error) {
-    res.statusCode = 500;
-    res.end("Internal Server Error");
-  }
-}
   
-module.exports = { gameRouteController, sendAllGames, sendUpdatedGames };
+module.exports = { sendAllGames, sendUpdatedGames };
