+++ utils/data/index.js
@@ -1,0 +1,3 @@
+const { readData, writeData } = require("./parse");
+
+module.exports = { readData, writeData };
